requirements.txt
python-dotenv
discord
pandas
